Refactor combine anonfile.py and pyapiwiz.py
